\name{aitr-package}
\alias{aitr-package}
\alias{aitr}
\docType{package}
\title{
\packageTitle{aitr}
}
\description{
\packageDescription{aitr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aitr}
\packageIndices{aitr}

}
\author{
\packageAuthor{aitr}

Maintainer: \packageMaintainer{aitr}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
