\name{itrrnr-package}
\alias{itrrnr-package}
\alias{itrrnr}
\docType{package}
\title{
\packageTitle{itrrnr}
}
\description{
\packageDescription{itrrnr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{itrrnr}
\packageIndices{itrrnr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{itrrnr}

Maintainer: \packageMaintainer{itrrnr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
